Removes v0, redirect downloads
Removes all v0 files. these endpoints now return 410 (gone), except for
the download action, which rewrites the url to return a redirect to the v1 api endpoint.

We may consider writing an nginx directive to handle this as it would be
far faster.
